+++ src/basemath/base4.c
@@ -1740,7 +1740,7 @@ idealpowred(GEN nf, GEN x, GEN n, long prec)
   if (typ(n) != t_INT) err(talker,"non-integral exponent in idealpowred");
   if (signe(n) == 0) return idealpow(nf,x,n);
   p1 = n+2; m = *p1;
-  y = x; j=1+bfffo(m); m<<=j; j = BITS_IN_LONG-j;
+  y = x; j=1+bfffo((ulong)m); m<<=j; j = BITS_IN_LONG-j;
   for (i=lgefint(n)-2;;)
   {
     for (; j; m<<=1,j--)
