+++ src/basemath/buch3.c
@@ -216,7 +216,7 @@ idealpowmodidele(GEN nf,GEN x,GEN n, GEN ideal,GEN sarch,GEN arch)
 
   if (gcmp1(n)) return x;
   p1 = n+2; m = *p1;
-  y=x; j=1+bfffo(m); m<<=j; j = BITS_IN_LONG-j;
+  y=x; j=1+bfffo((ulong)m); m<<=j; j = BITS_IN_LONG-j;
   for (i=lgefint(n)-2;;)
   {
     for (; j; m<<=1,j--)
