+++ src/basemath/polarit3.c
@@ -1117,7 +1117,7 @@ u_FpXQ_pow(GEN x, GEN n, GEN pol, ulong p)
   GEN p1 = n+2, y = x;
   long m,i,j;
   m = *p1;
-  j = 1+bfffo(m); m <<= j; j = BITS_IN_LONG-j;
+  j = 1+bfffo((ulong)m); m <<= j; j = BITS_IN_LONG-j;
   for (i=lgefint(n)-2;;)
   {
     for (; j; m<<=1,j--)
@@ -1159,7 +1159,7 @@ FpXQ_pow(GEN x, GEN n, GEN pol, GEN p)
   {
     av = avma;
     m = *p1; y = x;
-    j = 1+bfffo(m); m <<= j; j = BITS_IN_LONG-j;
+    j = 1+bfffo((ulong)m); m <<= j; j = BITS_IN_LONG-j;
     for (i=lgefint(n)-2;;)
     {
       for (; j; m<<=1,j--)
